+++ prp/models/phenotype.py
@@ -185,20 +185,6 @@ class SerotypeGene(GeneBase):
     """Container for serotype gene information"""
 
 
-class Shigatype(BaseModel):
-    """Container for shigatype gene information"""
-
-    rfb: Optional[str] = None
-    rfb_hits: Optional[float] = None
-    mlst: Optional[str] = None
-    flic: Optional[str] = None
-    crispr: Optional[str] = None
-    ipah: Optional[str] = None
-    predicted_serotype: Optional[str] = None
-    predicted_flex_serotype: Optional[str] = None
-    comments: Optional[str] = None
-
-
 class VirulenceGene(GeneBase, DatabaseReference):
     """Container for virulence gene information"""